requirements.txt
streamlit
numpy
pandas
matplotlib
plotly
openpyxl
python-docx